move core logic of convert_int_var_log_encoding to log::
